# revoker: Who doesn't love bit twiddling