Update the shadow bitmap in bulk, rather than with for-loops.  Running the test suite in sail shows ~7% speedup:

Before:

    Test runner: Allocator finished in 6771859 cycles
    Test runner: All tests finished in 7172781 cycles

After:

    Test runner: Allocator finished in 6251099 cycles
    Test runner: All tests finished in 6650597 cycles