@davidchisnall 2023-04-04
Results of the allocator benchmark on Sail:
```
Size    After      Before
32      76679518  82048087
64      42698710  48347087
128     24238618  29898029
256     15291611  20911973
512     10789531  13488733
1024    8348866   9663253
2048    7008814   7677549
4096    6160605   6440585
8192    5478284   5631514
16384   4938536   5015645
32768   4119331   4152346
65536   3902681   3918152
131072  3888274   3904099
```

I am quite surprised that we still see a speedup above 256 bytes.  Presumably the heap base is not aligned and so these still do some work painting the ends, so we are now doing 3 stores rather than the loop.  Above 16 KiB, the dela is pretty negligible, because painting the ends is such a small part of the total.